Add README for Anime MovieVerse project
Anime Movie Verse is an HTML project that showcases a collection of popular anime movies. It uses basic HTML elements to structure content, including headings, paragraphs, lists, links, and multimedia tags for images and trailers. This project is ideal for learning and practicing HTML fundamentals while exploring the world of anime cinema.